requirements.txt
streamlit
Pillow
google-generativeai
python-dotenv
requests
mistralai
PyPDF2
pydantic